split out memory example types vs global example types
