+++ src/provider-memory/exampleTypes.js
@@ -1,6 +1,6 @@
 let _ = require('lodash/fp')
 let F = require('futil-js')
-let strategies = require('./dataStrategies')
+let strategies = require('../dataStrategies')
 
 // like `_.includes` but casts everything to string first
 let toStringIncludes = _.curry((item, list) =>
@@ -119,28 +119,5 @@ module.exports = () => ({
         )
       ),
     }),
-  },
-  savedSearch: {
-    async filter(node, schema, { processGroup }) {
-      let debugSearch = x => processGroup(x, { debug: true })
-      let result = await strategies.analyzeTree(debugSearch, node.search, {
-        key: 'targetNode',
-      })
-      return result._meta.relevantFilters
-    },
-  },
-  subquery: {
-    filter: async (node, schema, { processGroup, getProvider }) =>
-      getProvider(node).types.facet.filter({
-        field: node.localField,
-        values: await strategies
-          .facet({
-            service: processGroup,
-            tree: node.search,
-            field: node.foreignField,
-            //size: 0 // <- put in once facet respects size: 0
-          })
-          .getNext(),
-      }),
   },
 })
